add tests, combine validate into contact form js template and css changes
