+++ fec/fec/static/scss/components/_contact-form.scss
@@ -52,4 +52,29 @@
       width: 100%;
     }
   }
+
+
+  input.invalid_border, select.invalid_border, textarea.invalid_border, checkbox.invalid_border, label.invalid_border {
+    border-color: $error !important; 
+  }
+
+
+  input:focus:invalid, select:focus:invalid, textarea:focus:invalid, label:focus:invalid {
+    outline: none;
+  }
+
+  span.error {
+    display:inline-block;
+    color: $error;
+  }
+
+  #checkbox_error {
+    display: block;
+  }
+
+
 }
+
+
+
+
